#6167 – Fix circular dependency, fix snake chain placement
